refactor(zmq): ZMQ connect mechanics move to the client layer
worker/worker.rs owned the entire ZMQ connect path — handshake-port
derivation (FNV-1a, pinned against serve.py), socket-address layout,
ipc-dir preparation, EOS resolution, and the engine handshake — while
reverse-importing ZmqEngineClient from routers::grpc to do it. The
worker layer was reaching into the router layer for its transport.

Move it all behind zmq_client::connect_for_worker: one connect path,
plain-reason errors, owned by the module that owns every other piece of
the ZMQ client (EOS, translate, streams). The worker keeps a thin
wrapper that maps into its client slot and WorkerError, and its reverse
import shrinks to the BackendClient/GrpcClient types it stores. The
pinned FNV vectors and address/ipc-dir tests move with the code. Also
corrects the stale EOS warning: connect-time ids being unavailable
falls back to the tokenizer EOS set at request time, it does not mean
generation runs to max_tokens.
